- fix dark mode still active - fix crash due to NumberFormatException in Home and Recipe

+++ app/src/main/java/com/irmwrs/recipeapp/fragments/RecipeListFragment.java
@@ -164,7 +164,7 @@ public class RecipeListFragment extends Fragment implements OnRecipeListener {
             if(recipe.isHighlighted){
                 highlightedList.add(recipe);
             }
-            if(Integer.parseInt(recipe.recipeRating) >= 4){
+            if(Double.parseDouble(recipe.recipeRating) >= 4){
                 topRatedList.add(recipe);
             }
         }
