+++ app/src/main/java/com/irmwrs/recipeapp/home/views/HomeFragment.java
@@ -182,7 +182,7 @@ public class HomeFragment extends Fragment implements OnRecipeListener {
             if(recipe.isHighlighted){
                 highlightedList.add(recipe);
             }
-            if(Integer.parseInt(recipe.recipeRating) >= 4){
+            if(Double.parseDouble(recipe.recipeRating) >= 4){
                 topRatedList.add(recipe);
             }
         }
